refactor(worker): isolate connector typing lifecycle
